Created the end- end floor, kitchen api for establishment area service

+++ establishment-area-service/src/main/java/com/teenthofabud/restaurant/solution/establishmentarea/kitchen/data/KitchenEntity.java
@@ -30,7 +30,7 @@ public class KitchenEntity extends TOABBaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private int kitchenId;
+	private Long kitchenId;
 
 	@Column(name = "name")
 	private String kitchenName;
